requirements.txt
boto3==1.17.73
botocore==1.20.73
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.25
protobuf==3.17.0
pycparser==2.20
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.4
Werkzeug==1.0.1